requirements.txt
Django>=5,<5.1
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.0,<6.0
psycopg2>=2.9.10,<2.10
drf_spectacular>=0.28,<0.29
eth-account>=0.13.4,<0.14
web3>=7.6.1,<7.7
redis>=5.2.1,<5.3
eth-utils>=5.1.0,<5.2
django-cors-headers>=4.3.1,<4.4
pillow>=11.0.0,<11.1
django-cleanup>=8.0.0,<8.1
randomname>=0.2.1,<0.3
python-dotenv>=1.0.1,<1.2
eth-abi>=5.0.1,<5.1
celery>=5.4.0,<5.5
django-celery-beat>=2.7.0,<2.8
watchdog>=6.0,<6.1
gunicorn>=22.0,<23.0
sentry-sdk>=2.22.0
